feat: Add DuckDB as optional storage backend (`do-memory-storage-duckdb`)
- Created new `do-memory-storage-duckdb` crate implementing the `StorageBackend` trait.
- Implemented columnar schema in `src/schema.rs` for episodes, patterns, embeddings, and monitoring.
- Integrated DuckDB VSS extension for indexed vector similarity search.
- Added analytical capabilities: session windowing, temporal decay scoring, and reward distribution analysis.
- Modularized implementation into `src/storage/` to comply with LOC limits.
- Fixed a regression in `memory-core` where retrieval cascade return types were incorrectly modified.
- Added integration tests covering monitoring metrics and raw analytics queries.
- Bypassed `rustup` cross-device link errors using stable toolchain environment variables.
