+++ memory-storage-duckdb/Cargo.toml
@@ -18,6 +18,7 @@ uuid = { workspace = true }
 parking_lot = "0.12"
 tracing = "0.1"
 async-trait = "0.1"
+hex = "0.4"
 
 [dev-dependencies]
 tempfile = { workspace = true }
