+++ methods_needed.txt
@@ -1,0 +1,32 @@
+cleanup_episodes_internal
+count_cleanup_candidates_internal
+delete_embedding_internal
+delete_episode_internal
+get_embedding_internal
+get_embeddings_batch_internal
+get_episode_internal
+get_heuristic_internal
+get_pattern_internal
+get_recommendation_feedback_internal
+get_recommendation_session_for_episode_internal
+get_recommendation_session_internal
+get_recommendation_stats_internal
+get_relationships_internal
+load_agent_metrics_internal
+load_execution_records_internal
+load_task_metrics_internal
+query_episodes_by_metadata_internal
+query_episodes_since_internal
+relationship_exists_internal
+remove_relationship_internal
+store_agent_metrics_internal
+store_embedding_internal
+store_embeddings_batch_internal
+store_episode_internal
+store_execution_record_internal
+store_heuristic_internal
+store_pattern_internal
+store_recommendation_feedback_internal
+store_recommendation_session_internal
+store_relationship_internal
+store_task_metrics_internal
